refactor(notice): extract shared webview notice widget
- Replace duplicated stateful WebView implementations across party and tuanwei notice pages with a single reusable WebViewNoticePage
- Simplify attachment processing in tuanwei_notice_beautify.js by replacing error-prone manual UTF-8 byte handling with standard URI utilities, and remove unnecessary try/catch blocks
- Add logic to strip link target attributes from all anchor tags to keep navigation within the WebView
- Reduce boilerplate code by consolidating shared WebView logic into the new reusable widget
